file rearrangement; update tokenization for dquotes
